refactor(suite-native): split portfolio content controls
